refactor(bufferThresholds): calcScroll → calcAxisScroll
